Food management + Custom times

+++ submit.php
@@ -14,7 +14,7 @@
 			echo "<div class='forms'>
 					<fieldset class='alert alert-success'>
 					<legend class='fat'>";
-            if(isset($_SESSION['food__quantity'])){
+            if(isset($_SESSION['food__quantity']) && !isset($_POST['fname']) && !isset($_POST['isCata'])){
 				if(isset($_SESSION['order_id'])){
 /**To edit an order, first need to DELETE previous order*/
 					$sql_del="DELETE FROM orders WHERE order_id={$_SESSION['order_id']}";
@@ -27,8 +27,13 @@
 				unset($_SESSION['cus_id']);
 				unset($_SESSION['food__quantity']);
 				session_destroy();
+				if(isset($_POST['time']) && !empty($_POST['time'])){
+					$time = "'{$_POST['time']}'";	
+				}else{
+					$time = 'curtime()';
+				}
                 $itemnum = count($food__quantity);
-                $sql_inserto = "INSERT orders(cus_id,date,time) VALUE($cus_id,curdate(),curtime())";
+                $sql_inserto = "INSERT orders(cus_id,date,time) VALUE($cus_id,curdate(),$time)";
                 $mysql->query($sql_inserto);
 				$order_id = mysql_insert_id();
                 for ($itemcount=0;$itemcount<$itemnum;$itemcount++) {
@@ -38,7 +43,7 @@
                     $mysql->query($sql_insertf);
                 }
                 echo "Create Order Successfully";  
-                header("refresh:1;url='index.php?page=current_orders'");				
+                header("refresh:1;url='index.php?page=current_orders'");		
             }else if(isset($_POST['fname'])){
 /**chaeck info and create a new customer*/				
 				if(!empty(preg_replace("/\s/","",(string)$_POST['fname']))){
@@ -49,7 +54,29 @@
 				}else{
 					echo "<script> history.back(-1)</script>";
 				}
-            }
+            }else if(isset($_POST['isCata'])){
+				$foodname = preg_replace("/\s/","",(string)$_POST['foodName']);
+				$foodPrice = preg_replace("/\s/","",(string)$_POST['price']);
+				if(!empty($foodname)){
+					if($_POST['isCata']=='food'){
+						$foodCata = $_POST['foodCata'];
+						if(isset($_POST['origId'])){
+							$foodId = $_POST['origId'];
+							$sql_newfood = "UPDATE food_catalogue SET cata_name = '$foodname',Price = $foodPrice,catalog_id = $foodCata WHERE food_id = $foodId";
+						}else{
+							$sql_newfood = "INSERT food_catalogue (cata_name,Price,catalog_id) VALUES ('$foodname',$foodPrice,$foodCata)";
+						}
+					}else if($_POST['isCata']=='cata'){
+						$cataId = $_POST['cataId'];
+						$sql_newfood = "INSERT food_catalogue (cata_name,catalog_id) VALUES ('$foodname',$cataId)";
+					}
+					$mysql->query($sql_newfood);
+					echo "Add New Food Successfully";
+					header("refresh:1;url='index.php?page=food&action=detail'");
+				}else{
+					echo "Wrong!<script>history.go(-1);</script>";
+				}
+			}
 			echo "		</legend>
 						<p>Back to Home Page in 1 seconds...</p>
 						<a href='index.php'>Back to Homepage immdiately</a>
